requirements.txt
annotated-types
anyio
asgiref

beautifulsoup4
cachetools
certifi
charset-normalizer
colorama
distro
Django
exceptiongroup

greenlet
grpcio
grpcio-status
h11
httpcore
httpx
idna
jiter
langchain
neo4j
numpy
openai
proto-plus
protobuf
psycopg2
pyasn1
pyasn1-modules
pydantic
pydantic-core
python-dotenv
pytz
PyYAML
requests
rsa
sniffio
soupsieve
SQLAlchemy
sqlparse
tqdm
typing-extensions
tzdata
urllib3
